Move TournamentErrors out; add document metadata
Extract TournamentErrors from Tournament.cs into a new TournamentErrors.cs file with appropriate namespace and usings, cleaning up the unused using in Tournament.cs. Update DocumentErrors.DocumentNotFound to return a structured Error with metadata (DocumentName) instead of an interpolated description string for better error payloads.
